==> notebooks/01_exploracao_dados.ipynb <==
import pandas as pd
import sys
import os


project_root = os.path.abspath(os.path.join('..'))
src_path = os.path.join(project_root, 'src')

if src_path not in sys.path:
    sys.path.append(src_path)
# -----------------------------------------------

from data.make_dataset import get_historical_data

print("Importação do módulo de dados bem-sucedida.")

# --- Vamos executar o "Driver" ---
# Você pode mudar a data de início se quiser
df_dados_brutos = get_historical_data(dat_start_filter='2024-06-01')

if df_dados_brutos is not None:
    print("\n--- Teste no Notebook Concluído ---")
    print(f"Total de {len(df_dados_brutos)} linhas carregadas.")
    
    # Mostra os 5 primeiros clientes
    print(df_dados_brutos['id_cliente'].unique()[:5])
    
    # Mostra um resumo dos dados
    print(df_dados_brutos.head())
